Cleaned up models directory by moving some modules to the utils directory. Updated the docs accordingly.
