+++ django_dicom/models/series.py
@@ -20,10 +20,10 @@ from django.contrib.postgres.fields import ArrayField, JSONField
 from django.core.validators import MinValueValidator
 from django.db import models
 from django.urls import reverse
-from django_dicom.models import help_text
 from django_dicom.models.dicom_entity import DicomEntity
-from django_dicom.models.fields import ChoiceArrayField
-from django_dicom.models.validators import digits_and_dots_only
+from django_dicom.models.utils.fields import ChoiceArrayField
+from django_dicom.models.utils import help_text
+from django_dicom.models.utils.validators import digits_and_dots_only
 from pathlib import Path
