+++ django_dicom/models/study.py
@@ -9,7 +9,7 @@ import logging
 from django.db import models
 from django.urls import reverse
 from django_dicom.models.dicom_entity import DicomEntity
-from django_dicom.models.validators import digits_and_dots_only
+from django_dicom.models.utils.validators import digits_and_dots_only
 
 
 class Study(DicomEntity):
