+++ django_dicom/models/utils/fields.py
