+++ django_dicom/models/values/unique_identifier.py
@@ -1,6 +1,6 @@
 from django.db import models
 from django_dicom.models.values.data_element_value import DataElementValue
-from django_dicom.models.validators import digits_and_dots_only
+from django_dicom.models.utils.validators import digits_and_dots_only
 
 
 class UniqueIdentifier(DataElementValue):
