+++ docs/source/modules/django_dicom.models.rst
@@ -6,7 +6,6 @@ Module contents
 
 .. automodule:: django_dicom.models
    :members:
-   :undoc-members:
    :show-inheritance:
 
 Subpackages
@@ -27,7 +26,6 @@ django\_dicom.models.data\_element module
 
 .. automodule:: django_dicom.models.data_element
    :members:
-   :undoc-members:
    :show-inheritance:
 
 django\_dicom.models.data\_element\_definition module
@@ -35,15 +33,6 @@ django\_dicom.models.data\_element\_definition module
 
 .. automodule:: django_dicom.models.data_element_definition
    :members:
-   :undoc-members:
-   :show-inheritance:
-
-django\_dicom.models.datetime module
-------------------------------------
-
-.. automodule:: django_dicom.models.datetime
-   :members:
-   :undoc-members:
    :show-inheritance:
 
 django\_dicom.models.dicom\_entity module
@@ -51,15 +40,6 @@ django\_dicom.models.dicom\_entity module
 
 .. automodule:: django_dicom.models.dicom_entity
    :members:
-   :undoc-members:
-   :show-inheritance:
-
-django\_dicom.models.fields module
-----------------------------------
-
-.. automodule:: django_dicom.models.fields
-   :members:
-   :undoc-members:
    :show-inheritance:
 
 django\_dicom.models.header module
@@ -67,31 +47,14 @@ django\_dicom.models.header module
 
 .. automodule:: django_dicom.models.header
    :members:
-   :undoc-members:
    :show-inheritance:
 
-django\_dicom.models.help\_text module
---------------------------------------
-
-.. automodule:: django_dicom.models.help_text
-   :members:
-   :undoc-members:
-   :show-inheritance:
 
 django\_dicom.models.image module
 ---------------------------------
 
 .. automodule:: django_dicom.models.image
    :members:
-   :undoc-members:
-   :show-inheritance:
-
-django\_dicom.models.logs module
---------------------------------
-
-.. automodule:: django_dicom.models.logs
-   :members:
-   :undoc-members:
    :show-inheritance:
 
 django\_dicom.models.patient module
@@ -99,7 +62,6 @@ django\_dicom.models.patient module
 
 .. automodule:: django_dicom.models.patient
    :members:
-   :undoc-members:
    :show-inheritance:
 
 django\_dicom.models.series module
@@ -107,7 +69,6 @@ django\_dicom.models.series module
 
 .. automodule:: django_dicom.models.series
    :members:
-   :undoc-members:
    :show-inheritance:
 
 django\_dicom.models.study module
@@ -115,13 +76,4 @@ django\_dicom.models.study module
 
 .. automodule:: django_dicom.models.study
    :members:
-   :undoc-members:
-   :show-inheritance:
-
-django\_dicom.models.validators module
---------------------------------------
-
-.. automodule:: django_dicom.models.validators
-   :members:
-   :undoc-members:
    :show-inheritance:
