+++ docs/source/modules/django_dicom.models.utils.rst
@@ -1,15 +1,42 @@
 django\_dicom.models.utils package
 ==================================
 
+Module contents
+---------------
+
+.. automodule:: django_dicom.models.utils
+   :members:
+   :show-inheritance:
+
 Submodules
 ----------
+
+django\_dicom.models.utils.fields module
+----------------------------------------
+
+.. automodule:: django_dicom.models.utils.fields
+   :members:
+   :show-inheritance:
+
+django\_dicom.models.utils.help\_text module
+--------------------------------------------
+
+.. automodule:: django_dicom.models.utils.help_text
+   :members:
+   :show-inheritance:
+
+django\_dicom.models.utils.logs module
+--------------------------------------
+
+.. automodule:: django_dicom.models.utils.logs
+   :members:
+   :show-inheritance:
 
 django\_dicom.models.utils.meta module
 --------------------------------------
 
 .. automodule:: django_dicom.models.utils.meta
    :members:
-   :undoc-members:
    :show-inheritance:
 
 django\_dicom.models.utils.utils module
@@ -17,14 +44,11 @@ django\_dicom.models.utils.utils module
 
 .. automodule:: django_dicom.models.utils.utils
    :members:
-   :undoc-members:
    :show-inheritance:
 
+django\_dicom.models.utils.validators module
+--------------------------------------------
 
-Module contents
----------------
-
-.. automodule:: django_dicom.models.utils
+.. automodule:: django_dicom.models.utils.validators
    :members:
-   :undoc-members:
    :show-inheritance:
